move completion under system command
